pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notebooklm-agent"
version = "0.1.0"
description = "Zero-token AI agent powered by Google NotebookLM"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "DEE", email = "dee@aissac.com" }]
keywords = ["notebooklm", "ai-agent", "zero-token", "gemini", "google"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "notebooklm-py>=0.4.0",
    "httpx>=0.27",
    "python-telegram-bot>=21.0",
    "click>=8.0",
    "rich>=13.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
nlm-agent = "notebooklm_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/notebooklm_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"